Implement j.nio.file.PosixException
`scala.scalanative.nio.fs.UnixException` is currently implemented in
terms of Posix: `scalanative.posix.errno._`. We create PosixExeception
to make this usage visible and implement UnixException in terms of
PosixException.
